rename `WorkflowBox` to `WorkflowPanel`
